+++ CHANGELOG.md
@@ -9,3 +9,4 @@
 * Added LoadingSpinner component
 * Removed TagsInput. [React Select](http://jedwatson.github.io/react-select/) is a better alternative.
 * Changed css prefix from `rui-` to `re-`.
+* Added Button component.